requirements.txt
streamlit
fal-client
python-dotenv
requests
pillow
replicate